Rewrite DeltaCalculator context tests and fix bugs
